requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# HTTP client (async)
httpx==0.28.1

# Data validation
pydantic==2.10.4
pydantic-settings==2.7.0
email-validator==2.2.0

# Auth
python-jose[cryptography]==3.3.0
PyJWT==2.9.0
bcrypt==4.2.1

# Database
sqlalchemy==2.0.36

# Optional: Redis caching
redis==5.2.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
